+++ src/GEDI/GEDI.jl
@@ -11,7 +11,7 @@ mutable struct GEDI_Granule{product} <: Granule
 end
 GEDI_Granule(product, args...) = GEDI_Granule{product}(args...)
 
-function Base.copy(g::GEDI_Granule{product}) where product
+function Base.copy(g::GEDI_Granule{product}) where {product}
     GEDI_Granule(product, g.id, g.url, g.info)
 end
 
@@ -19,7 +19,8 @@ function info(g::GEDI_Granule)
     gedi_info(g.id)
 end
 
-"""Derive info based on file id.
+"""
+Derive info based on file id.
 
 The id is built up as follows, see section 2.4 in the user guide
 GEDI02_A_2019110014613_O01991_T04905_02_001_01.h5
@@ -36,5 +37,13 @@ function gedi_info(filename)
     end
     days = Day(parse(Int, datetime[5:7]) - 1)  # Stored as #days in year
     datetime = datetime[1:4] * "0101" * datetime[8:end]
-    (type = Symbol(type * "_" * name), date = DateTime(datetime, gedi_date_format) + days, orbit = parse(Int, orbit[2:end]), track = parse(Int, track[2:end]), ppds = parse(Int, ppds), version = parse(Int, version), revision = parse(Int, revision))
+    (
+        type = Symbol(type * "_" * name),
+        date = DateTime(datetime, gedi_date_format) + days,
+        orbit = parse(Int, orbit[2:end]),
+        track = parse(Int, track[2:end]),
+        ppds = parse(Int, ppds),
+        version = parse(Int, version),
+        revision = parse(Int, revision),
+    )
 end