+++ src/ICESat/GLAH06.jl
@@ -1,7 +1,7 @@
 """
     points(g::ICESat_Granule{:GLAH06})
 
-Retrieve the points for a given GLAH06 (Land Ice) granule as a list of namedtuples
+Retrieve the points for a given ICESat GLAH06 (Land Ice) granule as a list of namedtuples
 The names of the tuples are based on the following fields:
 
 | Variable           | Original Field                        | Description                                           | Units                   |
@@ -11,10 +11,10 @@ The names of the tuples are based on the following fields:
 | `height`           | `Data_40HZ/Elevation_Surfaces/d_elev` | + `Data_40HZ/Elevation_Corrections/d_satElevCorr`     | m above WGS84 ellipsoid |
 | `datetime`         | `Data_40HZ/DS_UTCTime_40`             | Precise time of aquisiton                             | date-time               |
 | `quality` [^1]     | `Data_40HZ/Quality/elev_use_flg`      | & `Data_40HZ/Quality/sigma_att_flg` = 0               |                         |
-|                    | & `Data_40HZ/Waveform/i_numPk` = 1    | & `Data_40HZ.Elevation_Corrections/d_satElevCorr` < 3 | 1 = high quality        |
+|                    | & `Data_40HZ/Waveform/i_numPk` = 1    | & `Data_40HZ/Elevation_Corrections/d_satElevCorr` < 3 | 1 = high quality        |
 | `height_reference` | `land_ice_segments/dem/dem_h`         | Height of the (best available) DEM                    | height above WGS84      |
 
-You can get the output in a `DataFrame` with `DataFrame(points(g)`.
+You can get the output in a `DataFrame` with `DataFrame(points(g))`.
 
 [^1]: Smith, B., Fricker, H. A., Gardner, A. S., Medley, B., Nilsson, J., Paolo, F. S., ... & Zwally, H. J. (2020). Pervasive ice sheet mass loss reflects competing ocean and atmosphere processes. Science, 368(6496), 1239-1242.
 """
@@ -44,12 +44,7 @@ function points(granule::ICESat_Granule{:GLAH06}; step = 1)
 
         datetime = unix2datetime.(datetime .+ j2000_offset)
 
-        # convert from TOPEX/POSEIDON to WGS84 ellipsoid using Proj.jl
-        # This pipeline was validated against MATLAB's geodetic2ecef -> ecef2geodetic
-        pipe = Proj.proj_create(
-            "+proj=pipeline +step +proj=unitconvert +xy_in=deg +z_in=m +xy_out=rad +z_out=m +step +inv +proj=longlat +a=6378136.3 +rf=298.257 +e=0.08181922146 +step +proj=cart +a=6378136.3 +rf=298.257 +step +inv +proj=cart +ellps=WGS84 +step +proj=unitconvert +xy_in=rad +z_in=m +xy_out=deg +z_out=m +step +proj=axisswap +order=2,1",
-        )
-
+        pipe = topex_to_wgs84_ellipsoid()
         pts = Proj.proj_trans.(pipe, Proj.PJ_FWD, zip(longitude, latitude, height_ref))
         height_ref = [x[3] for x in pts]::Vector{Float64}
 
@@ -69,7 +64,7 @@ function points(granule::ICESat_Granule{:GLAH06}; step = 1)
                       (sigma_att_flg .== 0) .&
                       (i_numPk .== 1) .&
                       (saturation_correction .< 3),
-            height_ref = height_ref,
+            height_reference = height_ref,
         )
         return gt
     end