+++ src/ICESat/ICESat.jl
@@ -12,7 +12,7 @@ end
 ICESat_Granule(product, args...) = ICESat_Granule{product}(args...)
 
 function Base.copy(g::ICESat_Granule{product}) where {product}
-    return ICESat_Granule(product, g.id, g.url, g.info)
+    return ICESat_Granule(product, g.id, copy(g.url), copy(g.info))
 end
 
 function bounds(granule::ICESat_Granule)
@@ -32,7 +32,8 @@ end
 
 Base.isfile(g::ICESat_Granule) = Base.isfile(g.url)
 
-"""Derive info based on file id.
+"""
+Derive info based on file id.
 
 The id is built up as follows, see 1.2.5 in the user guide
 ATL03_[yyyymmdd][hhmmss]_[ttttccss]_[vvv_rr].h5
@@ -54,5 +55,13 @@ function icesat_info(filename)
         segment = parse(Int, segment),
         version = parse(Int, version),
         revision = parse(Int, revision),
+    )
+end
+
+function topex_to_wgs84_ellipsoid()
+    # convert from TOPEX/POSEIDON to WGS84 ellipsoid using Proj.jl
+    # This pipeline was validated against MATLAB's geodetic2ecef -> ecef2geodetic
+    pipe = Proj.proj_create(
+        "+proj=pipeline +step +proj=unitconvert +xy_in=deg +z_in=m +xy_out=rad +z_out=m +step +inv +proj=longlat +a=6378136.3 +rf=298.257 +e=0.08181922146 +step +proj=cart +a=6378136.3 +rf=298.257 +step +inv +proj=cart +ellps=WGS84 +step +proj=unitconvert +xy_in=rad +z_in=m +xy_out=deg +z_out=m +step +proj=axisswap +order=2,1",
     )
 end